requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.2.6
oauthlib==3.2.2
openai==1.82.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
pipenv==2025.0.2
platformdirs==4.3.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.32.0
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
tweepy==4.15.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
websocket-client==1.8.0
wsproto==1.2.0
